third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/object-src-param-url-blocked-expected.txt

CONSOLE ERROR: Refused to load plugin data from 'http://127.0.0.1:8080/plugins/resources/mock-plugin.pl?url' because it violates the following Content Security Policy directive: "object-src http://localhost:8080".

CONSOLE MESSAGE: PASS: Error occurred, so load was correctly blocked.
This test passes if there is a console message saying the plugin was blocked. 
